requirements.txt
Django>=4.2,<5.1
requests>=2.31.0
python-dotenv>=1.0.1
